Fix other picoquic threading issues
This removes the mutex from picoquic itself and instead uses a thread
safe queue of functions that the picoquic event loops runs. This allows
picoquic to be thread safe with function calls that other threads make to
picoquic.  This directly addresses CPU issues and problems with
stream creation from other threads.